Adjust GDI radar in fsnod06.map The radar was disabled after being captured by the player, which left a strange impression. Westwood clearly did this because it otherwise caused the player to put in a low power state as soon as the radar was captured, so the radar is now instead made to not drain power anymore.
It also turns out that GDI aircraft aren't responding to any of the villages being attacked in this mission, which is specifically caused by guard_mode_patch of ts-patches. Since this patch solves the issue where idle units automatically chase any enemy that passes by them all over the map, it's not reasonable to disable it, so hopefully the patch itself will be fixed.